[parser.c] Fix left childreen shift in expression node handling
The code fixes how we shift children nodes in an expression tree when moving
a node's childreen to the left. Previously, the right child node was
incorrectly set to the original right node, which led to incorrect expression
representation, and with that, an infinite loop.

When shifting children to the left, the new left child of the node becomes the
original left child, and the new right child should be the original right
node's left child.

Given an expression `50*E(20+120)`, shifting left results in `E(50*20)+120`.

Signed-off-by: Luis Silva <luismgsilva@outlook.com>