codegen: Add numerical values for global variables.
When a numerical value (e.g., an integer) is assigned to a
global variable, it will be placed in the ".data" section
of the assembly file.

    # C
    int a = 50;

    # Assembly
    section .data
    ; int a
    a: dd 50

Additionally, an "else" statement is added to handle cases where
the value type is not a string, ensuring it doesn't enter the
"asm_push" block twice.

Signed-off-by: Luis Silva <luismgsilva@outlook.com>